feat: refactor frontend theme package
* move frontend theme classes into their proper packages
